Support better interface for security advisories in Rust
Today, `web-bot-auth` messages cannot validate nonces. This changeset
adds that ability, and deprecates the `is_expired` and
`possibly_insecure` methods in favour of a simpler interface:
`verifier.get_details().possibly_insecure()`.

This interface accepts a closure that validates the nonce if able.
Because the component parametrs are available as `pub`, a user can
inspect other details as well for their benefit.