Expose the parsed label, rather than just the parameters, to the public (#40)
Systems need to be able to log aspects of the message that was received.

Previously, we only exposed `ParameterDetails`, which were sufficient
for logging a select few properties found in the message. However,
there's more value in exposing the contents of the chosen Signature and
Signature-Input label, allowing people to view the raw signature, as
well as the signed components in the message, so as to enforce their own
requirements.

This change drops exposing `ParameterDetails` directly and instead
exposes `ParsedLabel`, which both allows access to `ParameterDetails`
and additionally exposes the signature components and the signature
itself.